Reorganize utils into a utils folder
