+++ src/wallet/redux/reducers/running.ts
@@ -3,8 +3,9 @@ import * as states from '../../states';
 import * as actions from '../actions';
 import decode from '../../domain/decode';
 
-import { validSignature, validTransition, ourTurn, signPositionHex } from './utils';
 import { State } from 'fmg-core';
+import { ourTurn, validTransition } from '../../utils/reducer-utils';
+import { signPositionHex, validSignature } from '../../utils/signing-utils';
 
 export const runningReducer = (state: states.RunningState, action: actions.WalletAction): states.WalletState => {
   return waitForUpdateReducer(state, action);
