--- src/wallet/utils.ts
@@ -1,2 +1,0 @@
-
-export function unreachable(x: never) { return x; }
